django_app/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
confusable-homoglyphs==3.3.1
cssselect2==0.8.0
Django==5.2.2
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-form==3.4
django-bootstrap5==25.1
django-classy-tags==4.1.0
django-cms==5.0.1
django-cors-headers==4.7.0
django-entangled==0.6.2
django-filer==3.3.1
django-formtools==2.5.1
django-fsm==2.8.2
django-htmx==1.23.2
django-parler==2.3
django-polymorphic==4.1.0
django-registration==5.2.1
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.3
djangocms-attributes-field==4.0.0
djangocms-form-builder==0.3.2
djangocms-frontend==2.1.2
djangocms-link==5.0.1
djangocms-simple-admin-style==1.1.3
djangocms-text==0.8.8
djangocms-versioning==2.3.1
easy-thumbnails==2.10
idna==3.10
lxml==5.4.0
nh3==0.2.21
packaging==25.0
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.0
reportlab==4.4.1
requests==2.32.3
sqlparse==0.5.3
stripe==12.2.0
svglib==1.5.1
tinycss2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
